Our-Event-Server/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.4.0
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
djangorestframework==3.15.2
gunicorn==20.1.0
packaging==24.1
phonenumbers==8.13.47
pillow==10.4.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1